src/main/java/be/vdab/allesvoordekeuken/repositories/JpaArtikelRepository.java
package be.vdab.allesvoordekeuken.repositories;

import be.vdab.allesvoordekeuken.domain.Artikel;
import org.springframework.stereotype.Repository;

import javax.persistence.EntityManager;
import java.util.Optional;

@Repository
public class JpaArtikelRepository implements ArtikelRepository{
    private final EntityManager manager;

    public JpaArtikelRepository(EntityManager manager) {
        this.manager = manager;
    }


    @Override
    public Optional<Artikel> findByNummer(long id) {
        return Optional.ofNullable(manager.find(Artikel.class, id));
    }
}
